Major refactor, added function definitions.
The way we had our newlines set up was wrong because it didn't take into account
the significance they have in python. The "-> skip" allowed for invalid python
code, and since it was deeply integrated in the rules, I basically had to start
from scratch. Now the code is much more organized, scalable, and robust.